requirements.txt
aiogram==3.15.0
asyncpg==0.30.0
aportalsmp==1.0.0
python-dotenv==1.0.1